requirements.txt
numpy>=1.19.0
pillow>=8.0.0
tensorflow>=2.8.0
easyocr>=1.7.0
opencv-python>=4.5.0
matplotlib>=3.3.0
scikit-learn>=0.24.0